pyproject.toml
[project]
name = "coop"
version = "0.1.0"
description = ""
authors = [
    {name = "dongkang",email = "docryze@163.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain (>=1.1.0,<2.0.0)",
    "langchain-deepseek (>=1.0.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "rich (>=14.2.0,<15.0.0)"
]

[tool.poetry]
packages = [{include = "coop", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
